Rename python file to match class name.
