+++ network-config-analyzer/nca.py
@@ -8,7 +8,7 @@ import sys
 import time
 import os
 from SchemeRunner import SchemeRunner
-from ExecutionAssist import CLExecute
+from CLExecute import CLExecute
 from RESTServer import RestServer
